requirements.txt
Django~=6.0
django-phonenumber-field~=8.4.0
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.7.0
django-environ==0.11.2  # .env fayllar uchun (keyinroq foydali bo‘ladi)